Sema: Sink SanitizeExpr down into TypeCheckCodeCompletion.cpp
